# adapter/adapter.rest/adapter.rest.v2/src/main/webapp/api-docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "https://github.com/learning-layers/SocialSemanticServer/",
    "version" : "v2",
    "title" : "Social Semantic Server REST API",
    "termsOfService" : "https://github.com/learning-layers/",
    "contact" : {
      "email" : "dtheiler#tugraz$at"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/sss.adapter.rest.v2/rest",
...
  "paths" : {
...
    "/apps/apps/{apps}" : {
      "delete" : {
        "tags" : [ "apps" ],
        "summary" : "remove apps",
        "description" : "",
        "operationId" : "appsDelete",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "apps",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSAppsDeleteRet"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "SSAppsDeleteRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "apps" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
  }
}